+++ CHANGELOG.md
@@ -5,6 +5,10 @@ When adding entries, please treat them as if they could end up in a release any 
 
 Thank you!
 
+# Unreleased / next patch
+
+- Add an `@smithy4s.meta#unpackedOutput` trait that can be used to avoid packing an operation's output up in an additional case class in user code in [#1934](https://github.com/disneystreaming/smithy4s/pull/1934)
+
 # 0.19.0
 
 ## Binary-Compatible-Friendly Codegen Enabled in Core Module in [#1900](https://github.com/disneystreaming/smithy4s/pull/1900)